+++ src/routes/index.ts
@@ -14,6 +14,7 @@ import usersRoutes from './users.routes';
 import exploreRoutes from './explore.routes';
 import userSubscriptionRoutes from './user-subscription.routes';
 import notificationPreferencesRoutes from './notification-preferences.routes';
+import unsubscribeRoutes from './unsubscribe.routes';
 import notificationsRoutes from './notifications.routes';
 import deviceTokensRoutes from './device-tokens.routes';
 import reportsRoutes from './reports.routes';
@@ -45,6 +46,7 @@ v1.use('/user/notification-preferences', notificationPreferencesRoutes);
 v1.use('/user/notifications', notificationsRoutes);
 v1.use('/user/device-tokens', deviceTokensRoutes);
 v1.use('/reports', reportsRoutes);
+v1.use('/unsubscribe', unsubscribeRoutes);
 
 router.use('/v1', v1);